add shopping cart routes

+++ src/controllers/shoppingCart.js
@@ -1,5 +1,15 @@
 import { shoppingCartModel } from '../models';
+import shoppingCartRouter from '../routes/shoppingCart';
 
+
+export const listItems = async (req, res, next) => {
+  try {
+    const data = await shoppingCartModel.select('*');
+    res.status(200).json(data.rows);
+  } catch (error) {
+    next(error);
+  }
+};
 
 export const addItem = async (req, res, next) => {
   const { id, quantity } = req.body;
@@ -8,8 +18,8 @@ export const addItem = async (req, res, next) => {
 
   try {
     const data = await shoppingCartModel.addItem(columns, values);
-    const [ addedItem ] = data.rows;
-    res.status(200).json( addedItem );
+    const [addedItem] = data.rows;
+    res.status(200).json(addedItem);
   } catch (error) {
     next(error);
   }
@@ -21,29 +31,25 @@ export const updateItem = async (req, res, next) => {
   try {
     const data = await shoppingCartModel.updateItem(id, quantity);
     const [updatedItem] = data.rows;
-    res.status(200).json( updatedItem );
+    res.status(200).json(updatedItem);
   } catch (error) {
     next(error);
   }
 };
 
 export const removeItem = async (req, res, next) => {
-  const { id } = req.body;
+  const { id, clearEverything } = req.body;
 
   try {
-    const data = await shoppingCartModel.deleteItem(id);
+    let data;
+    if (clearEverything === undefined) {
+      data = await shoppingCartRouter.clear();
+    } else {
+      data = await shoppingCartModel.deleteItem(id);
+    }
     const { deletedItem } = data.rows;
-    res.status(200).json( deletedItem );
+    res.status(200).json(deletedItem);
   } catch(error) {
-    next(error);
-  }
-};
-
-export const eraseAllItems = async (req, res, next) => {
-  try {
-    await shoppingCartModel.clear();
-    res.status(204).end();
-  } catch (error) {
     next(error);
   }
 };
